Use common password widget
